Tidy up experiments
