+++ kifmm/Cargo.toml
@@ -57,22 +57,6 @@ csv = "1.3.0"
 num_cpus = "1.16.0"
 
 [[bench]]
-name = "precomputation_f32_m1"
-harness = false
-
-[[bench]]
-name = "precomputation_f64_m1"
-harness = false
-
-[[bench]]
-name = "precomputation_f32_amd"
-harness = false
-
-[[bench]]
-name = "precomputation_f64_amd"
-harness = false
-
-[[bench]]
 name = "direct_f32"
 harness = false
 
@@ -81,35 +65,19 @@ name = "direct_f64"
 harness = false
 
 [[bench]]
-name = "laplace_f32_m1"
+name = "laplace_mat_f64"
 harness = false
 
 [[bench]]
-name = "laplace_f32_m1_mat"
+name = "laplace_mat_f32"
 harness = false
 
 [[bench]]
-name = "laplace_f64_m1_mat"
+name = "laplace_f32"
 harness = false
 
 [[bench]]
-name = "laplace_f64_amd_mat"
-harness = false
-
-[[bench]]
-name = "laplace_f32_amd_mat"
-harness = false
-
-[[bench]]
-name = "laplace_f32_amd"
-harness = false
-
-[[bench]]
-name = "laplace_f64_m1"
-harness = false
-
-[[bench]]
-name = "laplace_f64_amd"
+name = "laplace_f64"
 harness = false
 
 [package.metadata.docs.rs]
