+++ kifmm/benches/laplace_mat_f64.rs
